+++ dart/lib/src/sentry_options.dart
@@ -7,6 +7,8 @@ import 'hub.dart';
 import 'protocol.dart';
 import 'utils.dart';
 
+const defaultEnvironment = 'production';
+
 /// Sentry SDK options
 class SentryOptions {
   /// Default Log level if not specified Default is DEBUG
@@ -98,7 +100,6 @@ class SentryOptions {
   /// Sets the release. SDK will try to automatically configure a release out of the box
   String release;
 
-// TODO: probably its part of environmentAttributes
   /// Sets the environment. This string is freeform and not set by default. A release can be
   /// associated with more than one environment to separate them in the UI Think staging vs prod or
   /// similar.
@@ -137,8 +138,14 @@ class SentryOptions {
   /// The server name used in the Sentry messages.
   String serverName;
 
+  Sdk _sdk = Sdk(name: sdkName, version: sdkVersion);
+
   /// Sdk object that contains the Sentry Client Name and its version
-  Sdk sdk;
+  Sdk get sdk => _sdk;
+
+  set sdk(Sdk sdk) {
+    _sdk = sdk ?? _sdk;
+  }
 
   // TODO: Scope observers, enableScopeSync